(feat) container frame and shutdowns (#20)
# Summary

This commit introduces the Frame TeaModel type and a new method on the
interactive container that can be used to shutdown the container. The
shutdown is useful for run follow-up work after shutting down.

## Notable Changes

- Small fix for collection item filtering. Specifying an ID will make
this more consistent.
- New `FrameView` that just wraps a bubbletea model into the interactive
container. Useful for run an existing program inside of the container
framework.
- New `Shutdown()` method on the interactive container that kills the
bubbletea program to free up the terminal output

## Change Type

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)